- Removed CDT dependency - Added Manifold dependency - Moved all gmm code to a legacy folder, to be deleted after reimplementation - Starting with a blank structure for the gmm - Parallelization cap for clang-tidy target
