+++ pycanha-core/include/pycanha-core/gmm/gmm.hpp
@@ -1,8 +1,4 @@
 #pragma once
 
-#include "pycanha-core/gmm/geometry.hpp"         // IWYU pragma: keep
-#include "pycanha-core/gmm/primitives.hpp"       // IWYU pragma: keep
-#include "pycanha-core/gmm/thermalmesh.hpp"      // IWYU pragma: keep
-#include "pycanha-core/gmm/transformations.hpp"  // IWYU pragma: keep
-#include "pycanha-core/gmm/triangulation.hpp"    // IWYU pragma: keep
-#include "pycanha-core/gmm/trimesh.hpp"          // IWYU pragma: keep
+#include "pycanha-core/gmm/geometrymodel.hpp"  // IWYU pragma: keep
+#include "pycanha-core/gmm/ids.hpp"            // IWYU pragma: keep
