+++ pycanha-core/include/pycanha-core/gmm/ids.hpp
@@ -1,0 +1,11 @@
+#pragma once
+
+#include <cstdint>
+
+namespace pycanha::gmm {
+
+enum class GeometryId : std::uint64_t {};
+enum class FaceId : std::uint64_t {};
+using NodeNum = std::int64_t;
+
+}  // namespace pycanha::gmm
