+++ pycanha-core/legacy/gmm/callbacks.hpp
