+++ pycanha-core/legacy/gmm/geometryutils.hpp
