+++ pycanha-core/legacy/gmm/test_primitives.cpp
