+++ pycanha-core/legacy/gmm/test_trimesh.cpp
