+++ pycanha-core/legacy/gmm/triangulation.cpp
