+++ pycanha-core/legacy/gmm/triangulation.hpp
